Microservice

+++ Backend/bus-service/src/main/java/com/capstone/bus_service/repository/BusScheduleRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.capstone.bus_service.entity.BusSchedule;
 
 public interface BusScheduleRepository extends JpaRepository<BusSchedule,Long> {
-	List<BusSchedule> findByBusId(int busId);
+	List<BusSchedule> findByBusId(long busId);
 }
